Added UpgradeDbHook
