+++ .gitignore
@@ -1,3 +1,4 @@
+*.bak
 *.class
 
 # Mobile Tools for Java (J2ME)
